requirements.txt
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
timm==0.9.8
numpy==1.23.5
pandas==2.2.1
scikit-learn==1.6.1
matplotlib==3.7.1
seaborn==0.12.2
Pillow==9.5.0
tensorflow==2.12.0
lime==0.2.0.1
scikit-image==0.22.0